// web/src/api/models/appChangesField.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * AppStoreCat API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */

export type AppChangesField = typeof AppChangesField[keyof typeof AppChangesField];


export const AppChangesField = {
  title: 'title',
  subtitle: 'subtitle',
  description: 'description',
  whats_new: 'whats_new',
  screenshots: 'screenshots',
  locale_removed: 'locale_removed',
} as const;